requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Django==3.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-multiselectfield==0.1.12
idna==2.9
numpy==1.18.1
Pillow==7.0.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
